Add support for local storage

+++ MyMovies/Models/Movie.swift
@@ -13,9 +13,8 @@ struct Movie: Decodable {
     let id: Int
     let originalTitle: String
     let posterPath: String?
-    let voteAverage: Float
+    let voteAverage: Double
     let overview: String
     let releaseDate: String
     var videos: [Video]?
-    
 }
